Rename parameter-or-variable-or-entitypage to country-model-item
