+++ src/components/pages/country-model-item.jsx
